requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tensorflow==2.15.0
keras==2.15.0
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
pillow==10.1.0
opencv-python==4.8.1.78
jupyter==1.0.0
ipykernel==6.27.1
plotly==5.17.0
streamlit==1.28.2
jinja2==3.1.2
python-multipart==0.0.6
aiofiles==23.2.1
locust==2.17.0
requests==2.31.0
boto3==1.34.0
python-dotenv==1.0.0
watchdog==3.0.0
psutil==5.9.6
prometheus-client==0.19.0